调整了meal_type的默认值，测试了新增餐品

+++ Fund/src/main/resources/MysqlCreate.sql
@@ -5,7 +5,7 @@ CREATE TABLE `fund` (
                         `fund_id` bigint NOT NULL COMMENT ' 余额id',
                         `account_id` bigint NOT NULL COMMENT ' 账户id',
                         `create_time` datetime NOT NULL default now() COMMENT '创建时间',
-                        `fund_amount` bigint NOT NULL DEFAULT 0 COMMENT '金额量',
+                        `fund_amount` double NOT NULL DEFAULT 0 COMMENT '金额量',
                         `is_delete` boolean NOT NULL DEFAULT false COMMENT '是否被删除',
                         PRIMARY KEY (`fund_id`),
                         UNIQUE KEY (`account_id`)
